Create a library for the integrators
This is almost like the previous commits.
The difference is that the part corresponding to the specialization
of the code is templated (and not exported).
So as to reuse this code, it has now be moved into a .t.cpp file.
The integration-*.cpp file now only contrains the definition required
by the library.