[Device] Correction of the unitTest of device to respect yaml config
Add an example of yaml file config that can be use for the device (the same generated by the unitTest with the function CreateYAMLFILE()).
Add two yaml files which reflect how the data are split in two in the sot, these ones are used by the unitTest. 
The control is set has constant = -0.5 for all joints in position; the device output a control signal respecting the position limits of the joints.